requirements.txt
Flask
mysql-connector-python
psycopg2-binary
python-dotenv
transformers
torch
sentencepiece
# accelerate